refactor: Refactored AdminSection styles and moved styles to separate file
