+++ src/styles/FoodAdmin/AdminSection.module.scss
@@ -122,11 +122,10 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  padding: 20px;
+  padding: 15px;
   border-radius: 12px;
   background-color: #f9f9f9;
   box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
-  transition: transform 0.2s ease, box-shadow 0.2s ease;
 
   @include responsive(mobile) {
     padding: 15px;
@@ -134,14 +133,9 @@
   }
 }
 
-.statusItem:hover {
-  transform: translateY(-5px);
-  box-shadow: 0 6px 18px rgba(0, 0, 0, 0.2);
-}
-
 .statusItem svg {
-  font-size: 2rem;
-  margin-bottom: 15px;
+  font-size: 1.8rem;
+  margin-bottom: 5px;
   color: #555;
 
   @include responsive(mobile) {
